refactor: seperate types
