+++ src/entities/types/entity-status.enum.ts
@@ -1,0 +1,4 @@
+export enum TrustEntityStatus {
+  ACTIVE = 'active',
+  REVOKED = 'revoked',
+}
