+++ src/entities/types/entity-type.enum.ts
@@ -1,0 +1,7 @@
+export enum TrustEntityType {
+  WALLET_PROVIDER = 'wallet_provider',
+  PID_PROVIDER = 'pid_provider',
+  ATTESTATION_PROVIDER = 'attestation_provider',
+  RP = 'rp',
+  REGISTRAR = 'registrar',
+}
