+++ src/entities/types/index.ts
@@ -1,0 +1,2 @@
+export * from './entity-type.enum';
+export * from './entity-status.enum';
